+++ packages/terra-grid/CHANGELOG.md
@@ -3,7 +3,16 @@ Changelog
 
 Unreleased
 ----------
+
+4.16.0 - (July 19, 2018)
+------------------
 ### Changed
+* Add .module extension to SCSS file(s)
+
+4.15.0 - (July 17, 2018)
+------------------
+### Changed
+* Minor dependency version bump
 * Formatting changes per eslint v4 update
 
 4.14.0 - (July 10, 2018)